fluxjob/requirements.txt
huggingface_hub
azure-storage-queue
azure-storage-blob